requirements.txt
torch
torchvision
tqdm
opencv_python
numpy
monai
pillow

